requirementsWithGpu.txt
-r requirements.txt
#note that tensorflow is broken as of 11/17/18 for python 3.7
tensorflow-gpu